requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-bcrypt==1.0.1

# Database & MongoDB
pymongo==4.6.0
dnspython==2.4.2

# Authentication & Security
bcrypt==4.0.1
itsdangerous==2.1.2

# Machine Learning & NLP
torch==2.3.1
torchvision==0.18.1
sentence-transformers==2.2.2
transformers==4.33.2
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.24.0

# Natural Language Processing
langdetect==1.0.9
deep-translator==1.11.4
beautifulsoup4==4.12.2
rapidfuzz==3.1.0

# Data Processing
openpyxl==3.1.2
requests==2.31.0

# Environment Management
python-dotenv==1.0.0

# NLP Utilities
nltk==3.9.1
sentencepiece==0.2.1
